moving polynomial_manifest to shared dir while Im at it
