+++ cpp/src/aztec/plonk/composer/standard_composer.hpp
@@ -2,7 +2,6 @@
 #include "composer_base.hpp"
 #include <transcript/manifest.hpp>
 #include <srs/reference_string/file_reference_string.hpp>
-#include "../proof_system/types/polynomial_manifest.hpp"
 
 namespace waffle {
 enum StandardSelectors { QM, QC, Q1, Q2, Q3, NUM };
