+++ cpp/src/aztec/plonk/proof_system/verifier/verifier.test.cpp
@@ -9,7 +9,6 @@
 #include <gtest/gtest.h>
 #include <srs/reference_string/file_reference_string.hpp>
 #include <polynomials/polynomial_arithmetic.hpp>
-#include <plonk/proof_system/types/polynomial_manifest.hpp>
 #include <plonk/proof_system/commitment_scheme/kate_commitment_scheme.hpp>
 
 namespace verifier_helpers {
