+++ cpp/src/aztec/plonk/proof_system/widgets/transition_widgets/transition_widget.hpp
@@ -6,7 +6,6 @@
 #include <unordered_map>
 
 #include <polynomials/iterate_over_domain.hpp>
-#include "../../types/polynomial_manifest.hpp"
 #include "../../types/prover_settings.hpp"
 #include "../../../../proof_system/proving_key/proving_key.hpp"
 #include "../../../../proof_system/work_queue/work_queue.hpp"
