+++ cpp/src/aztec/proof_system/polynomial_cache/polynomial_cache.hpp
@@ -3,7 +3,7 @@
 #include <unordered_map>
 #include <list>
 #include "polynomial_store.hpp"
-#include "../../plonk/proof_system/types/polynomial_manifest.hpp"
+#include "../../proof_system/types/polynomial_manifest.hpp"
 #include <algorithm>
 
 namespace waffle {
