+++ cpp/src/aztec/proof_system/verification_key/verification_key.hpp
@@ -3,7 +3,7 @@
 #include <srs/reference_string/reference_string.hpp>
 #include <polynomials/evaluation_domain.hpp>
 #include <crypto/sha256/sha256.hpp>
-#include "../../plonk/proof_system/types/polynomial_manifest.hpp"
+#include "../../proof_system/types/polynomial_manifest.hpp"
 
 namespace waffle {
